backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.2
chromadb==0.5.5
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.47.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# LLM Providers
anthropic==0.34.1
openai==1.47.0
google-generativeai==0.7.2
ollama==0.3.3

# MongoDB (optional — when DATABASE_PROVIDER=mongodb)
motor==3.5.1
pymongo==4.8.0

# Task queue (optional)
celery[redis]==5.4.0
